requirements.txt
# Core Django
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0

# Database
psycopg2-binary==2.9.9

# Async & Cache
celery==5.4.0
redis==5.0.8
django-redis==5.4.0
django-celery-results==2.6.0

# AI & ML
google-generativeai==0.8.3
sentence-transformers==3.0.1
chromadb==0.5.5

# File processing
PyPDF2==3.0.1
Pillow==10.4.0
pytesseract==0.3.11
python-docx==0.9.0

# Environment
python-decouple==3.8

# Monitoring & Observability
flower==2.0.1
sentry-sdk==2.13.0

# Development
django-extensions==4.2.0